+++ Packs/ServiceNow/Integrations/ServiceNowv2/test_data/response_constants.py
@@ -1891,3 +1891,12 @@ RESPONSE_GET_TABLE_NAME = {
         }
     ]
 }
+RESPONSE_DOCUMENT_ROUTE = {
+    "result": {
+        "display_name": "Incident: INC0000060",
+        "document_id": "document_id",
+        "document_table": "incident",
+        "queue": "queue_id",
+        "sys_id": "work_item_id"
+    }
+}