+++ Packs/ServiceNow/Integrations/ServiceNowv2/test_data/result_constants.py
@@ -269,3 +269,10 @@ EXPECTED_GET_TABLE_NAME = {
         }
     ]
 }
+EXPECTED_DOCUMENT_ROUTE = {
+    'ServiceNow.WorkItem(val.WorkItemID===obj.WorkItemID)':
+        {
+            'DisplayName': 'Incident: INC0000060', 'DocumentID': 'document_id', 'DocumentTable': 'incident',
+            'QueueID': 'queue_id', 'WorkItemID': 'work_item_id'
+        }
+}